+++ src/index.ts
@@ -7,15 +7,51 @@ This code is licensed under the BSD-3 license found in the LICENSE file in the r
 **/
 export { blueTheme as blue } from './blueTheme';
 export { blueDarkTheme as blueDark } from './blueDarkTheme';
-export { blueDarkAltTheme as blueDarkAlt } from './blueDarkAltTheme';
 
 // This will still need to be added to an end-user project if they want to access these properties without typescript yelling at them.
 declare global {
+    // eslint-disable-next-line @typescript-eslint/no-namespace
     namespace ReactNativePaper {
+        // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
         interface ThemeColors {
             primaryBase: string;
-            textSecondary: string;
+            primaryPalette: {
+                light: string;
+                main: string;
+                dark: string;
+            };
+            accentPalette: {
+                light: string;
+                main: string;
+                dark: string;
+            };
+            errorPalette: {
+                light: string;
+                main: string;
+                dark: string;
+            };
+            divider: string;
+            textPalette: {
+                primary: string;
+                secondary: string;
+                onPrimary: {
+                    light: string;
+                    main: string;
+                    dark: string;
+                };
+                highlight: string;
+                disabled: string;
+                hint: string;
+            };
+            actionPalette: {
+                hover: string;
+                active: string;
+                background: string;
+                disabled: string;
+                disabledBackground: string;
+            };
         }
+        // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
         interface ThemeFonts {
             bold: ThemeFont;
         }